Use original image in swirl effect when getting pixel value to swirl
I am not sure if this actually matters... BUT I have changed the getting
of the next pixel value to change to to use the original image instead
of the already manipulated image. I think this should avoid having weird
issues related to reusing pixel values and consistent values across the
board.