Create ImageProcessingContext
This commit creates the ImageProcessingContext to be passed into each of
the image processors. This means that we will do the inital computations
first (face detection, image instantiation), and then pass it to the
methods that can use whatever image context information they need. This
is to allow for multiple effects to be put on an image