Add photobooth server
This commit as a base "photobooth server" implementation. The purpose of
this server is to be always running, listening for keyboard input, and
start a "photobooth workflow" when a certain key is pressed. The
workflow is:

1) take pictures
2) put effects on pictures
3) combine pictures into one large image
4) print resulting image

For this initial commit, we do not actually take any pictures or print
anything, but instead use mock implementations for getting photos and
printing them (saving to disk).